# Added compression/decompression on chunk cache generation