--- a/Cargo.toml
+++ b/Cargo.toml
@@ -14,6 +14,7 @@ serde = "1.0.130"
 bincode = "1.3.3"
 futures-lite = "1.12.0"
 bevy_egui = "0.14.0"
+lz4_flex = "0.9.3"
 
 # Used by perf_counter
 once_cell = {version = "1.8.0", optional = true}